pass compile with -Werror=conversion

+++ examples/update.cpp
@@ -24,11 +24,11 @@ int main()
     std::vector<double> x, y;
 
     const double w = 0.05;
-    const double a = n/2;
+    const double a = (double)n/2.0;
 
     for (size_t i=0; i<n; i++) {
-        x.push_back(i);
-        y.push_back(a*sin(w*i));
+        x.push_back((double)i);
+        y.push_back(a*sin(w*(double)i));
     }
 
     std::vector<double> xt(2), yt(2);
